# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TerraLeb API
  version: 1.2.0
  description: >
    GIS field data collection API for projects, assignments, geospatial features,
    exports, notifications, and administration.
servers:
  - url: /
    description: Current TerraLeb API host (same origin as Swagger UI)
...
paths:
...
  /api/v1/projects:
    get:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: List projects accessible to current user
      responses:
        '200':
          description: Project list
        '401':
          description: Unauthorized
    post:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Create project (admin only)
      responses:
        '201':
          description: Project created
        '400':
          description: Validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/v1/projects/{projectId}:
    get:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Get project by ID
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
    put:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Update project (system admin or project admin)
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project updated
        '400':
          description: Validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Archive project
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project archived
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /api/v1/projects/{projectId}/stats:
    get:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Project statistics
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project statistics
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /api/v1/projects/{projectId}/features:
    get:
      tags: [Projects]
      security:
        - bearerAuth: []
      summary: Project features list
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project features
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /api/v1/projects/{projectId}/ai/readiness:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Check whether approved project data is ready for AI training
      description: Protected super-admin only. Read-only readiness probe. Does not run GEE, train models, classify, or write AI outputs.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - in: query
          name: label_field
          schema:
            type: string
        - in: query
          name: min_samples_per_class
          schema:
            type: integer
            minimum: 1
        - in: query
          name: scope_type
          schema:
            type: string
            enum: [project, governorate, district, city, custom_polygon, national]
      responses:
        '200':
          description: AI readiness result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiReadinessResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project not found
  /api/v1/projects/{projectId}/ai/published-layers:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: List published AI layers for Project Map overlays
      description: Viewer-safe endpoint. Users with project access can see only AI output layers with published status when project AI is enabled. Technical storage paths and unpublished review layers are not exposed.
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Published AI output layer list
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project not found
  /api/v1/projects/{projectId}/ai/published-predictions:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: List published AI prediction features for a project
      description: Viewer-safe endpoint backed by ai_prediction_feature. Returns the one published AI classification collection for the project map. Confidence is an attribute on each classification feature, not a separate map layer, and does not filter validation eligibility. AI predictions remain separate from spatial_feature until an admin approves/promotes a prediction.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - in: query
          name: detail
          schema:
            type: string
            enum: [overview, full]
            default: overview
        - in: query
          name: geometry
          schema:
            type: string
            enum: [aggregate, simplified, full]
            default: aggregate
        - in: query
          name: bounds
          schema:
            type: string
          description: Optional viewport bounds as west,south,east,north.
        - in: query
          name: zoom
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 2500
            default: 1800
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: class_label
          schema:
            type: string
        - in: query
          name: confidence_min
          schema:
            type: number
            minimum: 0
            maximum: 1
        - in: query
          name: confidence_max
          schema:
            type: number
            minimum: 0
            maximum: 1
        - in: query
          name: uncertainty_min
          schema:
            type: number
            minimum: 0
            maximum: 1
        - in: query
          name: uncertainty_max
          schema:
            type: number
            minimum: 0
            maximum: 1
      responses:
        '200':
          description: Published AI classification FeatureCollection payload from the database.
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project not found
  /api/v1/projects/{projectId}/ai/prediction-validation-tasks:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: List AI prediction validation tasks for a project
      description: Admin-only endpoint. Lists validation tasks backed by ai_prediction_feature rows. Assignment is optional; open tasks are available to project contributors. These tasks are not official field data and never approve or insert spatial_feature rows.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/AiPredictionValidationTaskStatus'
        - in: query
          name: assigned_to
          schema:
            type: string
            format: uuid
        - in: query
          name: ai_run_id
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Project AI prediction validation task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiPredictionValidationTaskListResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project not found
  /api/v1/projects/{projectId}/ai/prediction-validation-tasks/generate:
    post:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Generate AI prediction validation tasks
      description: Legacy admin-only task endpoint. Creates open validation tasks from classification ai_prediction_feature rows. Confidence is metadata, generation is idempotent for active tasks, and the task workflow does not write to spatial_feature.
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiPredictionValidationGenerateRequest'
      responses:
        '201':
          description: AI prediction validation tasks generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiPredictionValidationGenerateResponse'
        '400':
          description: Invalid generation request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project or selected AI layer not found
  /api/v1/projects/{projectId}/ai/settings:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Get project AI settings
      description: Protected super-admin only.
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: AI settings
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project not found
    put:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Save project AI settings
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiProjectSettingsRequest'
      responses:
        '200':
          description: AI settings saved
        '400':
          description: Validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project not found
    patch:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Patch project AI settings
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiProjectSettingsRequest'
      responses:
        '200':
          description: AI settings saved
        '400':
          description: Validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project not found
  /api/v1/projects/{projectId}/ai/governance:
    put:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Record or revoke separate AI training and publication authority
      description: Protected super-admin only. Enabling either purpose requires a documented basis and approval reference; ordinary AI enablement is insufficient.
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiGovernanceRequest'
      responses:
        '200':
          description: Governance authority record updated
        '403':
          description: Protected administrator required
        '422':
          description: Enabled purpose lacks a basis or approval reference
  /api/v1/projects/{projectId}/ai/runs:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: List project AI runs
      parameters:
        - $ref: '#/components/parameters/projectId'
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: status
          schema:
            type: string
            enum:
              [
                draft,
                queued,
                extracting_features,
                training,
                evaluating,
                classifying,
                ready_for_review,
                published,
                failed,
                cancelled,
              ]
      responses:
        '200':
          description: AI run list
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project not found
    post:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Create a draft or queued AI run record
      description: Creates an ai_run row and ai_run_log entry. A queued run is processed only when the backend AI worker command is running. Regional execution modes can run the controlled project-area pipeline and register review artifacts; national classification, publishing, Redis-backed queueing, and spatial_feature insertion are not started.
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiRunCreateRequest'
      responses:
        '201':
          description: AI run record created
        '400':
          description: Validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Readiness blockers prevent queueing
  /api/v1/projects/{projectId}/ai/runs/{runId}/publish:
    post:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Publish a completed AI run classification layer
      description: Admin endpoint. Publishes the run's classification ai_output_layer for project-map visibility and contributor validation. Publishing does not promote predictions into spatial_feature; promotion happens only after admin review approval.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/runId'
      responses:
        '200':
          description: AI run classification layer published
        '400':
          description: Run is not publishable
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project, run, or layer not found
        '409':
          description: Project AI is disabled or predictions are unavailable
  /api/v1/projects/{projectId}/ai/runs/{runId}/unpublish:
    post:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Unpublish a project AI run classification layer
      description: Admin endpoint. Removes contributor validation visibility for the published classification layer without deleting predictions or promoted spatial features.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/runId'
      responses:
        '200':
          description: AI run classification layer unpublished
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project, run, or layer not found
        '409':
          description: Run is not published
  /api/v1/projects/{projectId}/ai/runs/{runId}/validation-summary:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Get run-level AI prediction validation summary
      description: Counts all classification prediction features for the run regardless of confidence, plus contributor-validation and admin-review totals.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/runId'
      responses:
        '200':
          description: AI run validation summary
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project or run not found
  /api/v1/projects/{projectId}/ai/runs/{runId}/validations:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: List contributor validations for an AI run
      description: Admin endpoint. Lists contributor validation submissions for classification predictions in the run.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/runId'
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: AI run contributor validation list
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Project or run not found
  /api/v1/projects/{projectId}/ai/runs/{runId}/predictions/{predictionId}:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Get one AI prediction feature with validation state
      description: Returns the classification prediction, confidence attribute, contributor validation summary, the current user's validation if any, and admin-review state.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/runId'
        - $ref: '#/components/parameters/predictionId'
      responses:
        '200':
          description: AI prediction feature details
        '401':
          description: Unauthorized
        '403':
          description: Prediction is unpublished or user lacks access
        '404':
          description: Prediction not found
  /api/v1/projects/{projectId}/ai/predictions/{predictionId}/validations:
    post:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Submit contributor validation for an AI prediction feature
      description: Assigned contributors only. Each contributor can validate a published AI prediction once; many contributors can validate the same prediction.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/predictionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [validation_result]
              properties:
                validation_result:
                  type: string
                  enum: [correct, incorrect, unsure, cannot_verify]
                corrected_class:
                  type: string
                  nullable: true
                note:
                  type: string
                  nullable: true
                photo_media_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                gps_location:
                  type: object
                  nullable: true
                gps_accuracy_m:
                  type: number
                  nullable: true
      responses:
        '201':
          description: Contributor validation saved
        '400':
          description: Invalid validation payload
        '401':
          description: Unauthorized
        '403':
          description: User is not an assigned contributor or layer is unpublished
        '404':
          description: Prediction not found
        '409':
          description: Contributor already validated this feature or validation is closed
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: List contributor validations for one AI prediction feature
      description: Admin endpoint. Returns all contributor submissions for the prediction.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/predictionId'
      responses:
        '200':
          description: AI prediction contributor validation list
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Prediction not found
  /api/v1/projects/{projectId}/ai/predictions/{predictionId}/my-validation:
    get:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Get the current contributor's AI prediction validation
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/predictionId'
      responses:
        '200':
          description: Current user's validation or null
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Prediction not found
  /api/v1/projects/{projectId}/ai/predictions/{predictionId}/admin-review:
    post:
      tags: [AI]
      security:
        - bearerAuth: []
      summary: Admin-review an AI prediction feature
      description: Admin endpoint. Approving a prediction promotes it idempotently into spatial_feature with AI provenance and use_for_future_training=true; rejecting keeps the prediction for audit without promotion.
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/predictionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approval_status]
              properties:
                approval_status:
                  type: string
                  enum: [approved, rejected, needs_more_validation]
                approved_class:
                  type: string
                  nullable: true
                admin_note:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Admin review saved
        '400':
          description: Invalid review payload
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Prediction not found
        '409':
          description: Layer is unpublished or prediction cannot be reviewed
...
  /api/v1/assignments/project/{projectId}:
    get:
      tags: [Assignments]
      security:
        - bearerAuth: []
      summary: Get assignments for a project
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Assignment list
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/v1/projects/{projectId}/offline-package:
    get:
      tags: [Offline / Sync]
      security:
        - bearerAuth: []
      summary: Get the project-scoped offline package manifest
      description: Available only to contributors with an approved contributor assignment for the project.
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Project schema, workflow counts, base-map manifest, and package version
        '401':
          description: Unauthorized
        '403':
          description: Contributor assignment required
        '404':
          description: Project not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Paste a valid TerraLeb access token. Swagger UI adds the Bearer prefix.
    aiCallbackSecret:
      type: apiKey
      in: header
      name: x-ai-callback-secret
      description: Service-to-service callback credential used only by the trusted AI server.
    metricsToken:
      type: http
      scheme: bearer
      description: Operational metrics token when METRICS_TOKEN is configured.
  parameters:
    projectId:
      in: path
      name: projectId
      required: true
      schema:
        type: string
        format: uuid
...
    runId:
      in: path
      name: runId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    AiPredictionValidationTaskStatus:
      type: string
      enum: [open, assigned, in_progress, submitted, accepted, rejected, cancelled]
...
    AiPredictionValidationTaskListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/AiPredictionValidationTask'
            status_counts:
              type: object
              additionalProperties:
                type: integer
            no_spatial_feature_writes:
              type: boolean
            not_official_field_data:
              type: boolean
        pagination:
          type: object
...